+++ lib/index.js
@@ -4,6 +4,6 @@ import PrintService from './printService';
 import ConfigApi from './configApi';
 import QuickTravelApi from './quicktravelApi';
 import findPrinters from './printerMatcher';
-import printToPrinter from './quicketsApi';
+import printToPrinter from './quickPrintApi';
 
 export { PrintService, ConfigApi, QuickTravelApi, findPrinters, printToPrinter };
