+++ lib/printService.js
@@ -1,7 +1,7 @@
 import ConfigApi from './configApi';
 import QuickTravelApi from './quicktravelApi';
 import findPrinters from './printerMatcher';
-import printToPrinter from './quicketsApi';
+import printToPrinter from './quickPrintApi';
 
 class PrintService {
   constructor({ quicktravel, config }) {
