+++ lib/quickPrintApi.js
