+++ test/quickPrintApiTest.js
